pyproject.toml
[project]
name = "vibeos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13,<4",
    "python-telegram-bot>=21.0,<22.0",
    "pi_sdk @ git+https://github.com/garymjr/pi_sdk.git",
]
